requirements.txt
aiohttp==3.8.4 ; python_version >= "3.10" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.10" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.10" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
charset-normalizer==3.1.0 ; python_version >= "3.10" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.10" and python_version < "4.0"
dataclasses-json==0.5.7 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
elastic-transport==8.4.0 ; python_version >= "3.10" and python_version < "4"
elasticsearch==8.7.0 ; python_version >= "3.10" and python_version < "4"
email-validator==2.0.0.post2 ; python_version >= "3.10" and python_version < "4.0"
fastapi[all]==0.95.1 ; python_version >= "3.10" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.10" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.10" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httpcore==0.17.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
httpx==0.24.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
langchain==0.0.147 ; python_version >= "3.10" and python_version < "4.0"
lxml==4.9.2 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
marshmallow-enum==1.5.1 ; python_version >= "3.10" and python_version < "4.0"
marshmallow==3.19.0 ; python_version >= "3.10" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
numexpr==2.8.4 ; python_version >= "3.10" and python_version < "4.0"
numpy==1.24.3 ; python_version >= "3.10" and python_version < "4.0"
openai==0.27.4 ; python_version >= "3.10" and python_version < "4.0"
openapi-schema-pydantic==1.2.4 ; python_version >= "3.10" and python_version < "4.0"
orjson==3.8.10 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
regex==2023.3.23 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
soupsieve==2.4.1 ; python_version >= "3.10" and python_version < "4.0"
sqlalchemy==1.4.47 ; python_version >= "3.10" and python_version < "4.0"
starlette==0.26.1 ; python_version >= "3.10" and python_version < "4.0"
tenacity==8.2.2 ; python_version >= "3.10" and python_version < "4.0"
tiktoken==0.3.3 ; python_version >= "3.10" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.10" and python_version < "4.0"
typing-inspect==0.8.0 ; python_version >= "3.10" and python_version < "4.0"
ujson==5.7.0 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.10" and python_version < "4"
uvicorn[standard]==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
watchfiles==0.19.0 ; python_version >= "3.10" and python_version < "4.0"
websockets==11.0.2 ; python_version >= "3.10" and python_version < "4.0"
yarl==1.9.1 ; python_version >= "3.10" and python_version < "4.0"
pandas==2.0.2 ; python_version >= "3.10" and python_version < "4.0"
PyGithub==1.58.2 ; python_version >= "3.10" and python_version < "4.0"
